Use fbs fork with a custom pyinstaller hook to handle vlc binary dependencies and a custom hook to address the tcl/tk issue
The fbs build system lacks a clear solution for sourcing system dependencies or using the PyInstaller APIs that are designed for doing exactly that. The PyInstaller hooks api is especially useful for sourcing binary dependencies. It is both fast and intuitive. This commit uses a custom fork of fbs in which I have added a few custom hook using the hooks api.

tcl/tk hook - Solves PyInstaller issue ttps://github.com/pyinstaller/pyinstaller/issues/3753

libvlc hook - Bundles system libvlc binary dependencies. This is the most important one. It saves from having to do this manually for different platforms, and makes for a lighter build script.