Improve fullscreen support on macOS with multiple displays
Fixes issue reported by Anneka, where fullscreen mode would make other screens go black. The fix uses screen geometry and a frameless window instead of the fullscreen mode when on macOS. This means when the video is displayed fullscreen on the main OS screen (with the dock and top bar) the dock and top bar will be overlaid on top of the video. We should add a check for whether the screen is the main OS screen, and use the regular fullscreen mode if so.